+++ tests/test_modules/test_metadata_manager.py
@@ -1,0 +1,5 @@
+from mite_data.modules.metadata_manager import MetadataManager
+
+
+def test_init_valid():
+    assert isinstance(MetadataManager(), MetadataManager)