+++ CHANGELOG.md
@@ -17,7 +17,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Set the status of all entries to `active`
 - reworked project structure
 - complete rework of `mite_data` CLI
-- added MITE entry-corresponding FASTA files
+- added functionality to automatically generate metadata for MITE entries
+- added functionality to download protein FASTA files and build a BLAST DB
+- added functionality to update MITE entries to the newest version of `mite_schema`
 
 ## [1.1] 11-08-2024